MAINT: broaden allowed lists in `_DtypeLike`
Closes https://github.com/numpy/numpy-stubs/issues/42.

`np.dtype` accepts a wide variety of lists as dtypes, some examples
include:

- `[('R', 'u1'), ('G', np.unicode_, 1)]`
- `[('R', 'u1', (2, 2)), ('G', np.unicode_, 1)]`

Without further hints mypy is going to type those as `List[object]`,
so precise hints will require that users add annotations. In addition,
typing the above is challenging because the inputs have to be a list,
which is invariant, so you can't use `Union`s inside the list to
reduce the number of cases. Work around that by just accepting
`List[Any]`.